refactored home page into components
