+++ src/webapp/features/home/components/GameList/GameList.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import styles from "./home.module.css";
-import RecentGame from "./RecentGame";
-import { clearRecentGames } from "../../app/localStorage";
+import styles from "./GameList.module.css";
+import RecentGame from "../RecentGame/RecentGame";
+import { clearRecentGames } from "../../../..//app/localStorage";
 
 type GameListProps = {
   games: UserGameInfo[];
@@ -11,7 +11,7 @@ const TurnList = (props: GameListProps) => {
   const { games } = props;
 
   return (
-    <div>
+    <div className="RecentGameList">
       <h5>Recent games</h5>
       <ul className={styles.gameList}>
         {games.map((g) => {
