+++ src/webapp/features/home/components/RecentGame/RecentGame.module.css
@@ -1,0 +1,9 @@
+/*  begin experimental styling */
+
+.recentGame {
+  width: 250px;
+  height: 200px;
+  background-color: turquoise;
+}
+
+/* end experimental styling */
