+++ src/webapp/features/home/home.module.css
@@ -1,9 +1,3 @@
 .content {
   text-align: center;
 }
-
-.gameList {
-  list-style: none;
-  text-align: center;
-  padding: 0px;
-}
